Make User model, adding :username, :email and ..
:password using a migration. Run the migration and add validations for
presence for all 3 attributes and uniqueness for :username and :email. Generate
the Users controller with empty methods for #new and #create. Make Users a resource
in config/routes.rb with routes only for :new and :create. Create the view file
for :new with dummy text. Start a Rails server and go to http://localhost:3000/users/new to
see the dummy text displayed.